Don't add rawFileKey to models.Recording.userGetAttributes
The recordings/get/:id API was unintentionally modifying the global
model.Recording.userGetAttributes list. This was causing rawFileKey to
be added to unrelated queries. After enough calls to this API the list
of requested columns became so large the query column limit was
reached causing the DB to reject them.