/api/V1/users -> /api/V1/users/:username
Replace the API which returns all the users matching some criteria
with one which returns the details for just one username. The latter
is the all the web UI actually wants.

It's also good to avoid APIs which take raw sequelize strings as are
security risks as and limit our implementation flexibility.

The web UI is being updated to switch to this new API.