Store audio recordings in the Recordings table
The pre-existing AudioRecordings API still exists but it now interacts
with the Recordings table instead. The AudioRecordings API is now
marked as deprecated. Most of the API handler logic for dealing with
recordings has been moved out to api/V1/recordingUtil.js so that it
can be shared by the Recordings and AudioRecordings APIs.

A DB migration is included which imports all audio recordings into
Recordings.