Avoid double response when audio recording not found
When getFileData fails reponseUtil.invalidFileRequest is called and
then an error is thrown. The thrown error is then picked up by the
catch handled which then ends up calling
responseUtil.invalidDatapointGet, resulting in 2 HTTP responses being
sent (not valid).

This is a likely fix for one of the issues reported in #71.

This code is all going away soon but it's good to fix this in the meantime.