Docker based development workflow
The `docker-start.sh` shell script has been replaced with a Python
based `run` script. This supports both the case of running a "clean"
API server instances where the source tree is isolated in the
container and npm dependencies are installed from scratch, as well as
a developer friendly mode where the source tree is mapped into the
container and the server restarts whenever the code changes.

All services are now started and configured by the docker entry point
script using environment variables instead of having an external
script act inside the container. This cleans things up signficantly
and make the system easier to understand.

To simplify testing and development of cacophony-processing the
fileProcessing API port is now available outside.

Documentation has been updated to match, with much of the existing
docs removed. `Dockerfile` and `docker-entrypoint.sh` are now more
precise descriptions of what is required to run the API server and
these will stay current as the software evolves.

Closes #108.