Migrate to sequelize v4 - class/instance definitions
The main change is that classMethods and instanceMethods are removed from sequelize.define. Instead, they need to be declared as methods on the model. Class members are defined on the model, while instance methods are defined on the prototype (see http://docs.sequelizejs.com/manual/tutorial/upgrade-to-v4.html#config-options).

To implement this, I took the model definition from the return statement at the end to a const at the beginning. So the pattern in each model file is: declare name, attributes and options, then declare the model, then class methods on the model (starting with associations), then instance methods on the model prototype, then data arrays (public and local), then the model object is returned. Local functions used only within the model are declared as needed; validation functions are declared outside the scope of the exports function.

The other main change is to replace string-based operators, which are deprecated due to security issues, with symbol-based operators. This is done in two steps: 1) In this commit, in the model index, setting operatorsAliases to create aliases for the string-based operators in use. 2) Next commit will replace those operators with symbol-based operators in the models.