Initial work for #199, support both legacy and new version of the PATCH endpoint. Also fix #205, all fields are now properly validated.
Reworked the way Error constructors are handled to make use of new language features. This way is better in terms of output and is also more concise.

Removed catch block in middleware that didn't need to be there that I added in an earlier PR, can instead be caught in the global error handler.

Reworked the User hooks to handle password hashing as they were being double hashed with the rework of the validation.

Fixed response code for invalid JSON to return a 422 rather than a 500.

Added tests for user PATCH and POST to prevent regression. Added middleware tests to check some status codes are as expected.