Support legacy "animal" and "event" tag fields.
To support the transition between the legacy and new recording tag
fields the API will now accept and return "animal" and "event". During
tag creation "animal" is translated to "what" and "event" is
translated to "detail". For output both the old and new fields are
provided.

This translation layer can be removed once all clients have been
updated to use the new fields.